service/requirements.txt
python-dotenv
schedule
pymongo
selenium
webdriver_manager
paramiko
scp
easyocr
opencv-python
paramiko
puresnmp